[MISC] move include/seqan3/{core,utility}/simd/simd_traits.hpp
