+++ include/seqan3/alignment/pairwise/detail/policy_optimum_tracker_simd.hpp
@@ -18,7 +18,7 @@
 #include <seqan3/alignment/pairwise/detail/policy_optimum_tracker.hpp>
 #include <seqan3/core/simd/simd.hpp>
 #include <seqan3/utility/simd/algorithm.hpp>
-#include <seqan3/core/simd/simd_traits.hpp>
+#include <seqan3/utility/simd/simd_traits.hpp>
 #include <seqan3/range/views/zip.hpp>
 #include <seqan3/utility/type_traits/lazy_conditional.hpp>
