+++ include/seqan3/alignment/pairwise/policy/simd_affine_gap_policy.hpp
@@ -23,7 +23,7 @@
 #include <seqan3/core/configuration/configuration.hpp>
 #include <seqan3/utility/simd/concept.hpp>
 #include <seqan3/utility/simd/algorithm.hpp>
-#include <seqan3/core/simd/simd_traits.hpp>
+#include <seqan3/utility/simd/simd_traits.hpp>
 
 namespace seqan3::detail
 {
