+++ include/seqan3/core/simd/detail/simd_algorithm_avx2.hpp
@@ -17,7 +17,7 @@
 #include <seqan3/utility/simd/concept.hpp>
 #include <seqan3/core/simd/detail/builtin_simd_intrinsics.hpp>
 #include <seqan3/core/simd/detail/builtin_simd.hpp>
-#include <seqan3/core/simd/simd_traits.hpp>
+#include <seqan3/utility/simd/simd_traits.hpp>
 
 //-----------------------------------------------------------------------------
 // forward declare avx2 simd algorithms that use avx2 intrinsics
