+++ include/seqan3/core/simd/view_iota_simd.hpp
@@ -17,7 +17,7 @@
 #include <seqan3/core/concept/core_language.hpp>
 #include <seqan3/utility/simd/concept.hpp>
 #include <seqan3/utility/simd/algorithm.hpp>
-#include <seqan3/core/simd/simd_traits.hpp>
+#include <seqan3/utility/simd/simd_traits.hpp>
 
 namespace seqan3::detail
 {
