+++ test/performance/utility/simd/simd_algorithm_benchmark.cpp
@@ -11,7 +11,7 @@
 #include <benchmark/benchmark.h>
 
 #include <seqan3/utility/simd/algorithm.hpp>
-#include <seqan3/core/simd/simd_traits.hpp>
+#include <seqan3/utility/simd/simd_traits.hpp>
 #include <seqan3/core/simd/simd.hpp>
 
 // ----------------------------------------------------------------------------
